requirements.txt
python-dotenv>=1.0.0
openai>=1.3.0
google-generativeai>=0.3.0
httpx>=0.24.0
langchain>=0.0.200
langchain-openai>=0.0.2
cachetools>=5.3.0
pytz>=2023.3
aiohttp>=3.8.1
uuid>=1.30
requests>=2.31.0
tenacity>=8.2.0
async-timeout>=4.0.0
dataclasses>=0.6
typing-extensions>=4.0.0
python-dateutil>=2.8.2
tzlocal>=4.0.0
google-auth-oauthlib>=0.4.6
google-auth>=2.6.0
google-api-python-client>=2.0.0
cachetools>=4.2.4
tenacity>=8.0.1
python-dotenv>=0.19.0
pytz>=2021.3

